backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
Pillow==10.2.0
geopy==2.4.1
python-dateutil==2.8.2
email-validator==2.1.0
